+++ .github/workflows/main.yml
@@ -541,6 +541,7 @@ jobs:
             cargo check -p wasmtime --no-default-features --features runtime,gc,component-model &&
             cargo check -p cranelift-control --no-default-features &&
             cargo check -p pulley-interpreter --features encode,decode,disas,interp
+            cargo check -p wasmtime-wasi-io --no-default-features
         # Use `cross` for illumos to have a C compiler/linker available.
         - target: x86_64-unknown-illumos
           os: ubuntu-latest
