+++ crates/wasi-io/src/lib.rs
@@ -1,7 +1,10 @@
 pub mod bindings;
-mod io;
+mod impls;
 pub mod poll;
 pub mod streams;
 mod view;
 
 pub use view::{IoImpl, IoView};
+
+#[doc(no_inline)]
+pub use async_trait::async_trait;
