+++ crates/wasi-io/src/poll.rs
@@ -1,11 +1,7 @@
-use crate::bindings::wasi::io::poll;
-use crate::view::{IoImpl, IoView};
-use anyhow::{anyhow, Result};
+use anyhow::Result;
 use std::any::Any;
-use std::collections::HashMap;
 use std::future::Future;
 use std::pin::Pin;
-use std::task::{Context, Poll};
 use wasmtime::component::{Resource, ResourceTable};
 
 pub type PollableFuture<'a> = Pin<Box<dyn Future<Output = ()> + Send + 'a>>;
@@ -19,9 +15,9 @@ pub type ClosureFuture = Box<dyn Fn() -> PollableFuture<'static> + Send + 'stati
 /// or writable. So, rather than containing a Future, which can only become Ready once, a
 /// Pollable contains a way to create a Future in each call to `poll`.
 pub struct Pollable {
-    index: u32,
-    make_future: MakeFuture,
-    remove_index_on_delete: Option<fn(&mut ResourceTable, u32) -> Result<()>>,
+    pub(crate) index: u32,
+    pub(crate) make_future: MakeFuture,
+    pub(crate) remove_index_on_delete: Option<fn(&mut ResourceTable, u32) -> Result<()>>,
 }
 
 /// A trait used internally within a [`Pollable`] to create a `pollable`
@@ -42,8 +38,11 @@ pub struct Pollable {
 /// parameters.
 ///
 /// ```
+/// # // stub out so we don't need a dep to build the doctests:
+/// # mod tokio { pub mod time { pub use std::time::{Duration, Instant}; pub async fn sleep_until(_:
+/// Instant) {} } }
 /// use tokio::time::{self, Duration, Instant};
-/// use wasmtime_wasi::{IoView, Subscribe, subscribe, Pollable, async_trait};
+/// use wasmtime_wasi_io::{IoView, poll::{Subscribe, subscribe, Pollable}, async_trait};
 /// use wasmtime::component::Resource;
 /// use wasmtime::Result;
 ///
@@ -115,95 +114,4 @@ where
     };
 
     Ok(table.push_child(pollable, &resource)?)
-}
-
-impl<T> poll::Host for IoImpl<T>
-where
-    T: IoView,
-{
-    async fn poll(&mut self, pollables: Vec<Resource<Pollable>>) -> Result<Vec<u32>> {
-        type ReadylistIndex = u32;
-
-        if pollables.is_empty() {
-            return Err(anyhow!("empty poll list"));
-        }
-
-        let table = self.table();
-
-        let mut table_futures: HashMap<u32, (MakeFuture, Vec<ReadylistIndex>)> = HashMap::new();
-
-        for (ix, p) in pollables.iter().enumerate() {
-            let ix: u32 = ix.try_into()?;
-
-            let pollable = table.get(p)?;
-            let (_, list) = table_futures
-                .entry(pollable.index)
-                .or_insert((pollable.make_future, Vec::new()));
-            list.push(ix);
-        }
-
-        let mut futures: Vec<(PollableFuture<'_>, Vec<ReadylistIndex>)> = Vec::new();
-        for (entry, (make_future, readylist_indices)) in table.iter_entries(table_futures) {
-            let entry = entry?;
-            futures.push((make_future(entry), readylist_indices));
-        }
-
-        struct PollList<'a> {
-            futures: Vec<(PollableFuture<'a>, Vec<ReadylistIndex>)>,
-        }
-        impl<'a> Future for PollList<'a> {
-            type Output = Vec<u32>;
-
-            fn poll(mut self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<Self::Output> {
-                let mut any_ready = false;
-                let mut results = Vec::new();
-                for (fut, readylist_indicies) in self.futures.iter_mut() {
-                    match fut.as_mut().poll(cx) {
-                        Poll::Ready(()) => {
-                            results.extend_from_slice(readylist_indicies);
-                            any_ready = true;
-                        }
-                        Poll::Pending => {}
-                    }
-                }
-                if any_ready {
-                    Poll::Ready(results)
-                } else {
-                    Poll::Pending
-                }
-            }
-        }
-
-        Ok(PollList { futures }.await)
-    }
-}
-
-impl<T> crate::bindings::wasi::io::poll::HostPollable for IoImpl<T>
-where
-    T: IoView,
-{
-    async fn block(&mut self, pollable: Resource<Pollable>) -> Result<()> {
-        let table = self.table();
-        let pollable = table.get(&pollable)?;
-        let ready = (pollable.make_future)(table.get_any_mut(pollable.index)?);
-        ready.await;
-        Ok(())
-    }
-    async fn ready(&mut self, pollable: Resource<Pollable>) -> Result<bool> {
-        let table = self.table();
-        let pollable = table.get(&pollable)?;
-        let ready = (pollable.make_future)(table.get_any_mut(pollable.index)?);
-        futures::pin_mut!(ready);
-        Ok(matches!(
-            futures::future::poll_immediate(ready).await,
-            Some(())
-        ))
-    }
-    fn drop(&mut self, pollable: Resource<Pollable>) -> Result<()> {
-        let pollable = self.table().delete(pollable)?;
-        if let Some(delete) = pollable.remove_index_on_delete {
-            delete(self.table(), pollable.index)?;
-        }
-        Ok(())
-    }
 }
